backend/outdated/conftest.py
import pytest
from pytest_factoryboy import register
from rest_framework.test import APIClient

from .outdated import factories

register(factories.DependencyFactory)
register(factories.DependencyVersionFactory)
register(factories.ProjectFactory)


@pytest.fixture
def client(db):
    """Return rest framework client, includes db."""
    return APIClient()


@pytest.fixture
def date_format():
    """Return the format for dates, used when converting strings to a datetime object."""
    return "%Y-%m-%d"
